break out data into its own data files
