+++ my-react-portfolio/package.json
@@ -13,6 +13,7 @@
     "framer-motion": "^11.11.11",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-scroll": "^1.9.0",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
