+++ my-react-portfolio/src/components/Header.jsx
@@ -1,10 +1,17 @@
 import styled from "styled-components";
 import { TextBlock } from "./UI-components/TextBlock";
 import { TextBlockSection } from "./UI-components/TextBlock";
+import { motion } from "framer-motion";
 
 export const StyledH2 = styled.h2`
   color: ${(props) => (props.white ? "#ffffff" : "#000000")};
 `;
+
+const images = [
+  "https://images.unsplash.com/photo-1729731322011-f945437445be?q=80&w=2667&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  "https://images.unsplash.com/photo-1728815235820-3a894b2f2807?q=80&w=3240&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  "https://images.unsplash.com/photo-1728327511854-fb04fa8fb2e4?q=80&w=3464&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+];
 
 export const Header = () => {
   return (
@@ -19,17 +26,42 @@ export const Header = () => {
       </TextBlockSection>
 
       <div className="images">
-        <img
-          src="https://images.unsplash.com/photo-1729731322011-f945437445be?q=80&w=2667&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
+        <motion.img
+          src={images[0]}
           alt=""
+          initial={{ opacity: 0 }}
+          animate={{
+            opacity: 1,
+            rotate: -6,
+            x: 170,
+            y: 35,
+          }}
+          transition={{ duration: 0.8, delay: 0.2 }}
         />
-        <img
-          src="https://images.unsplash.com/photo-1728815235820-3a894b2f2807?q=80&w=3240&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
+        <motion.img
+          src={images[1]}
           alt=""
+          initial={{ opacity: 0 }}
+          animate={{
+            opacity: 1,
+            rotate: 0,
+            x: 0,
+            y: 0,
+            zIndex: 1,
+          }}
+          transition={{ duration: 0.8, delay: 0.4 }}
         />
-        <img
-          src="https://images.unsplash.com/photo-1728327511854-fb04fa8fb2e4?q=80&w=3464&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
+        <motion.img
+          src={images[2]}
           alt=""
+          initial={{ opacity: 0 }}
+          animate={{
+            opacity: 1,
+            rotate: 6,
+            x: -170,
+            y: 35,
+          }}
+          transition={{ duration: 0.8, delay: 0.6 }}
         />
       </div>
       <TextBlock
