+++ my-react-portfolio/src/data/articleData.js
@@ -1,0 +1,20 @@
+export const ArticleData = [
+  {
+    title: "How to stop being scared: Learning to love code",
+    description:
+      "The chat bot app is a conversational AI-powered tool designed to enhance user experience by providing instant, personalized, and automated responses to user inquiries.",
+    src: "https://images.unsplash.com/photo-1728815235820-3a894b2f2807?q=80&w=3240&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  },
+  {
+    title: "How to stop being scared: Learning to love code",
+    description:
+      "The chat bot app is a conversational AI-powered tool designed to enhance user experience by providing instant, personalized, and automated responses to user inquiries.",
+    src: "https://images.unsplash.com/photo-1728815235820-3a894b2f2807?q=80&w=3240&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  },
+  {
+    title: "How to stop being scared: Learning to love code",
+    description:
+      "The chat bot app is a conversational AI-powered tool designed to enhance user experience by providing instant, personalized, and automated responses to user inquiries.",
+    src: "https://images.unsplash.com/photo-1728815235820-3a894b2f2807?q=80&w=3240&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  },
+];
